Added new tool MESSI. The tool is applyed to spaci al transcriptomics data.
Mixture of Experts for Spacial Signaling genes Identification (MESSI), is a python framework which applies the Mixture of Experts machine learning technique to model interactions between cells from spatial single cell transcriptomics data. For each cell type in the data, the MESSI model uses as input a set of intra- and inter- signaling genes to predict the expression of a set of response genes.

https://www.scrna-tools.org/tools#MESSI